[refactor] 배송 경로 기록 서비스 관련 파일 위치 이동 및 변수명 변경
