+++ delivery/src/main/java/com/sparta/rooibus/delivery/application/service/DeliveryLogServiceImpl.java
@@ -1,18 +1,16 @@
 package com.sparta.rooibus.delivery.application.service;
 
 import com.sparta.rooibus.delivery.application.dto.request.SearchRequest;
-import com.sparta.rooibus.delivery.application.dto.response.SearchDeliveryResponse;
-import com.sparta.rooibus.delivery.domain.entity.Delivery;
 import com.sparta.rooibus.delivery.domain.entity.DeliveryLog;
 import com.sparta.rooibus.delivery.domain.model.Pagination;
 import com.sparta.rooibus.delivery.domain.repository.DeliveryLogRepository;
-import com.sparta.rooibus.delivery.presentation.controller.CreateDeliveryLogRequest;
-import com.sparta.rooibus.delivery.presentation.controller.CreateDeliveryLogResponse;
-import com.sparta.rooibus.delivery.presentation.controller.DeleteDeliveryLogResponse;
-import com.sparta.rooibus.delivery.presentation.controller.GetDeliveryLogResponse;
-import com.sparta.rooibus.delivery.presentation.controller.SearchDeliveryLogResponse;
-import com.sparta.rooibus.delivery.presentation.controller.UpdateDeliveryLogRequest;
-import com.sparta.rooibus.delivery.presentation.controller.UpdateDeliveryLogResponse;
+import com.sparta.rooibus.delivery.application.dto.request.CreateDeliveryLogRequest;
+import com.sparta.rooibus.delivery.application.dto.response.CreateDeliveryLogResponse;
+import com.sparta.rooibus.delivery.application.dto.response.DeleteDeliveryLogResponse;
+import com.sparta.rooibus.delivery.application.dto.response.GetDeliveryLogResponse;
+import com.sparta.rooibus.delivery.application.dto.response.SearchDeliveryLogResponse;
+import com.sparta.rooibus.delivery.application.dto.request.UpdateDeliveryLogRequest;
+import com.sparta.rooibus.delivery.application.dto.response.UpdateDeliveryLogResponse;
 import jakarta.persistence.EntityNotFoundException;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
